move ServiceAction to new package and modify a definition to be more descriptive
